requirements.txt
PyGithub==1.54.1
PyYAML==5.3.1
tqdm==4.56.0
